Add CONTAINER authentication method
This method is based on the systemd nspawn container system.
When authenticating from the container, the hostname is used as ID
and the machine-id is used as shared secret. The server provides
a challenge, the client provides another challenge and a hash is
computed from these inputs. If the server can match the client
hash, the authentication is successful.